pyproject.toml
[project]
name = "fastapi-onnx-sklearn"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sogo Kato", email = "sogo.kato@outlook.com" }
]
dependencies = [
    "scikit-learn>=1.5.2",
    "numpy>=2.2.0",
    "skl2onnx>=1.17.0",
    "onnxruntime>=1.20.1",
    "fastapi>=0.115.6",
    "pydantic-settings>=2.6.1",
    "uvicorn>=0.32.1",
    "pillow>=11.0.0",
    "python-multipart>=0.0.19",
    "img2feat>=0.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_onnx_sklearn"]
